refactor: SourcemapBuilder
