+++ crates/oxc_codegen/src/lib.rs
@@ -12,19 +12,20 @@ mod context;
 mod gen;
 mod gen_ts;
 mod operator;
+mod sourcemap_builder;
 
-use std::{str::from_utf8_unchecked, vec};
+use std::str::from_utf8_unchecked;
 
 #[allow(clippy::wildcard_imports)]
 use oxc_ast::ast::*;
 use oxc_span::{Atom, Span};
 use oxc_syntax::{
-    identifier::{is_identifier_part, LS, PS},
+    identifier::is_identifier_part,
     operator::{BinaryOperator, UnaryOperator, UpdateOperator},
     precedence::Precedence,
     symbol::SymbolId,
 };
-use sourcemap::{SourceMap, SourceMapBuilder};
+use sourcemap_builder::SourcemapBuilder;
 
 pub use crate::{
     context::Context,
@@ -32,13 +33,6 @@ pub use crate::{
     operator::Operator,
 };
 // use crate::mangler::Mangler;
-
-#[derive(Debug)]
-pub struct LineOffsetTable {
-    columns: Option<Vec<usize>>,
-    byte_offset_to_first: usize,
-    byte_offset_to_start_of_line: usize,
-}
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct CodegenOptions {
@@ -71,14 +65,7 @@ pub struct Codegen<const MINIFY: bool> {
     indentation: u8,
 
     // sourcemap
-    enable_sourcemap: bool,
-    source_id: u32,
-    last_generated_update: usize,
-    last_position: Option<u32>,
-    line_offset_tables: Vec<LineOffsetTable>,
-    sourcemap_builder: SourceMapBuilder,
-    generated_line: u32,
-    generated_column: u32,
+    sourcemap_builder: SourcemapBuilder,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -106,14 +93,7 @@ impl<const MINIFY: bool> Codegen<MINIFY> {
             start_of_arrow_expr: 0,
             start_of_default_export: 0,
             indentation: 0,
-            enable_sourcemap: false,
-            source_id: 0,
-            last_generated_update: 0,
-            last_position: None,
-            line_offset_tables: vec![],
-            sourcemap_builder: SourceMapBuilder::new(None),
-            generated_line: 0,
-            generated_column: 0,
+            sourcemap_builder: SourcemapBuilder::default(),
         }
     }
 
@@ -121,30 +101,25 @@ impl<const MINIFY: bool> Codegen<MINIFY> {
     // self.mangler = Some(mangler);
     // }
 
-    pub fn build(mut self, program: &Program<'_>) -> String {
-        program.gen(&mut self, Context::default());
+    pub fn with_sourcemap(&mut self, source: &str, source_name: &str) -> &mut Self {
+        self.sourcemap_builder
+            .with_enable_sourcemap(true)
+            .with_source_and_name(source, source_name);
+        self
+    }
+
+    pub fn build(&mut self, program: &Program<'_>) -> String {
+        program.gen(self, Context::default());
         self.into_code()
     }
 
-    pub fn into_code(self) -> String {
+    pub fn into_code(&mut self) -> String {
         // SAFETY: criteria of `from_utf8_unchecked`.are met.
-        unsafe { String::from_utf8_unchecked(self.code) }
+        unsafe { String::from_utf8_unchecked(std::mem::take(&mut self.code)) }
     }
 
-    pub fn build_with_sourcemap(
-        mut self,
-        program: &Program<'_>,
-        source: &str,
-        source_name: &str,
-    ) -> (String, SourceMap) {
-        self.enable_sourcemap = true;
-        self.line_offset_tables = Self::generate_line_offset_tables(source);
-        self.source_id = self.sourcemap_builder.add_source(source_name);
-        self.sourcemap_builder.set_source_contents(self.source_id, Some(source));
-        program.gen(&mut self, Context::default());
-        let code = std::mem::take(&mut self.code);
-        // SAFETY: criteria of `from_utf8_unchecked`.are met.
-        (unsafe { String::from_utf8_unchecked(code) }, self.sourcemap_builder.into_sourcemap())
+    pub fn into_sourcemap(self) -> sourcemap::SourceMap {
+        self.sourcemap_builder.into_sourcemap()
     }
 
     fn code(&self) -> &Vec<u8> {
@@ -430,151 +405,11 @@ impl<const MINIFY: bool> Codegen<MINIFY> {
     }
 
     fn add_source_mapping(&mut self, position: u32) {
-        self.internal_add_source_mapping(position, None);
+        self.sourcemap_builder.add_source_mapping(&self.code, position, None);
     }
 
     fn add_source_mapping_for_name(&mut self, position: u32, name: &str) {
-        self.internal_add_source_mapping(position, Some(name));
-    }
-
-    fn internal_add_source_mapping(&mut self, position: u32, name: Option<&str>) {
-        if self.enable_sourcemap {
-            if matches!(self.last_position, Some(last_position) if last_position == position) {
-                return;
-            }
-            let (original_line, original_column) = self.search_original_line_and_column(position);
-            self.update_generated_line_and_column();
-            // println!("span {}, {:?} => {:?}", position, (original_line, original_column), (self.generated_line, self.generated_column));
-            let name_id = name.map(|s| self.sourcemap_builder.add_name(s));
-            self.sourcemap_builder.add_raw(
-                self.generated_line,
-                self.generated_column,
-                original_line,
-                original_column,
-                Some(self.source_id),
-                name_id,
-            );
-            self.last_position = Some(position);
-        }
-    }
-
-    #[allow(clippy::cast_possible_truncation)]
-    fn search_original_line_and_column(&self, position: u32) -> (u32, u32) {
-        let result = self
-            .line_offset_tables
-            .partition_point(|table| table.byte_offset_to_start_of_line <= position as usize);
-        let original_line = if result > 0 { result - 1 } else { 0 };
-        let line = &self.line_offset_tables[original_line];
-        let mut original_column = (position as usize) - line.byte_offset_to_start_of_line;
-        if original_column >= line.byte_offset_to_first {
-            if let Some(cols) = &line.columns {
-                original_column = cols[original_column - line.byte_offset_to_first];
-            }
-        }
-        (original_line as u32, original_column as u32)
-    }
-
-    fn update_generated_line_and_column(&mut self) {
-        // SAFETY: criteria of `from_utf8_unchecked`.are met.
-        let s = unsafe { std::str::from_utf8_unchecked(&self.code[self.last_generated_update..]) };
-        // println!("\n---- {:?}", s);
-        for (i, ch) in s.chars().enumerate() {
-            match ch {
-                '\r' | '\n' | LS | PS => {
-                    // Handle Windows-specific "\r\n" newlines
-                    if ch == '\r' && self.code[self.last_generated_update + i + 1] == b'\n' {
-                        continue;
-                    }
-                    self.generated_line += 1;
-                    self.generated_column = 0;
-                }
-                _ => {
-                    if ch as u32 <= 0xFF {
-                        self.generated_column += 1;
-                    } else {
-                        self.generated_column += 2;
-                    }
-                }
-            }
-        }
-        self.last_generated_update = self.code.len();
-    }
-
-    fn generate_line_offset_tables(content: &str) -> Vec<LineOffsetTable> {
-        let mut tables = vec![];
-        let mut columns = None;
-        let mut column = 0;
-        let mut column_byte_offset = 0;
-        let mut line_byte_offset = 0;
-        let mut byte_offset_to_first = 0;
-        for (i, ch) in content.chars().enumerate() {
-            // Mark the start of the next line
-            if column == 0 {
-                line_byte_offset = i;
-            }
-
-            // Start the mapping if this character is non-ASCII
-            if !ch.is_ascii() && columns.is_none() {
-                column_byte_offset = i - line_byte_offset;
-                byte_offset_to_first = column_byte_offset;
-                columns = Some(vec![]);
-            }
-
-            // Update the per-byte column offsets
-            if let Some(columns) = &mut columns {
-                for _ in column_byte_offset..(i - line_byte_offset) {
-                    columns.push(column);
-                }
-            }
-
-            match ch {
-                '\r' | '\n' | LS | PS => {
-                    // Handle Windows-specific "\r\n" newlines
-                    if ch == '\r' && content.chars().nth(i + 1) == Some('\n') {
-                        column += 1;
-                        continue;
-                    }
-
-                    tables.push(LineOffsetTable {
-                        columns,
-                        byte_offset_to_first,
-                        byte_offset_to_start_of_line: line_byte_offset,
-                    });
-                    column = 0;
-                    columns = None;
-                    byte_offset_to_first = 0;
-                    column_byte_offset = 0;
-                }
-                _ => {
-                    if ch as u32 <= 0xFF {
-                        column += 1;
-                    } else {
-                        column += 2;
-                    }
-                }
-            }
-        }
-        // Mark the start of the next line
-        if column == 0 {
-            line_byte_offset = content.len();
-        }
-
-        // Do one last update for the column at the end of the file
-        if let Some(columns) = &mut columns {
-            for _ in column_byte_offset..(content.len() - line_byte_offset) {
-                columns.push(column);
-            }
-        }
-
-        tables.push(LineOffsetTable {
-            columns,
-            byte_offset_to_first,
-            byte_offset_to_start_of_line: line_byte_offset,
-        });
-
-        // dbg!(&tables);
-
-        tables
+        self.sourcemap_builder.add_source_mapping(&self.code, position, Some(name));
     }
 }
